Fixed array error in json

+++ src/main/java/au/com/numbrcrunchr/domain/EquityCalculator.java
@@ -2,8 +2,8 @@ package au.com.numbrcrunchr.domain;
 
 // TODO Create test for EquityCalculator
 public class EquityCalculator {
-	public long calculateEquity(long marketValue, long loanBalance) {
-		long equity = (marketValue * 80 / 100) - loanBalance;
+	public double calculateEquity(double marketValue, double loanBalance) {
+		double equity = (marketValue * 80 / 100) - loanBalance;
 		return equity < 0 ? 0 : equity;
 	}
 }
