+++ src/main/java/au/com/numbrcrunchr/domain/LoanAmortisationScheduleCalculator.java
@@ -27,7 +27,7 @@ public class LoanAmortisationScheduleCalculator implements Serializable {
      * @return
      */
     private List<Amortisation> calculatePrincipalAndInterestAmortisationSchedule(
-            long loanAmount, int termInYears, double interestRate) {
+    		double loanAmount, int termInYears, double interestRate) {
         double loanBalance = loanAmount;
         double r = 1 + (interestRate / 100) / 12;
         int termInMonths = termInYears * 12;
@@ -65,7 +65,7 @@ public class LoanAmortisationScheduleCalculator implements Serializable {
         this.interestCalculator = interestCalculator;
     }
 
-    public AmortisationSchedule calculateAmortisationSchedule(long loanAmount,
+    public AmortisationSchedule calculateAmortisationSchedule(double loanAmount,
             int interestOnlyPeriod, int termInYears, double interestRate) {
         List<Amortisation> interestOnlyAmortisations = new ArrayList<Amortisation>(
                 interestOnlyPeriod);
@@ -87,7 +87,7 @@ public class LoanAmortisationScheduleCalculator implements Serializable {
         return AmortisationSchedule.createAmortisationSchedule(amortisations);
     }
 
-    public AmortisationSchedule calculateAmortisationSchedule(long loanAmount,
+    public AmortisationSchedule calculateAmortisationSchedule(double loanAmount,
             int termInYears, double interestRate) {
         return this.calculateAmortisationSchedule(loanAmount, 0, termInYears,
                 interestRate);
